df-server/pyproject.toml
[project]
name = "df-server"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "fastapi[standard]>=0.112.1",
    "sqlmodel>=0.0.21",
    "xdg-base-dirs>=6.0.1",
    "ruamel-yaml>=0.18.6",
    "jinja2>=3.1.4",
    "df_modules",
    "gpxpy>=1.6.2",
]
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/df_server"]
